Update eth ticker example.
Eth ticker is now split to a proxy and client part.
